// src/app/enums/api-path.enum.ts
export enum APIPath {
  OFFER_LETTER_REQUEST = "/offerletters/offerletter/",
  USER_ROLE = "/users/get_users_by_role/?user_role=",
  JOB_ASSIGNMENT = "/jobdescriptions/jobassignment/",
  JOB_SUBMISSION = "/jobdescriptions/jobsubmission/",
  JOB_DESCRIPTION = "/jobdescriptions/jobdescription/",
  CAMPAIGN_DASHBOARD = "/schedulers/send_clean_data/",
  JOB_DESCRIPTION_NOTES = "/jobdescriptions/jobdescriptionnotes/",
  USERS = "/users/users/",
  COUNTRIES = "/users/countries/",
  CANDIDATE_TABLE = "/reports/candidate_table",
  CANDIDATE_CSV = "/reports/candidate_csv",
  CANDIDATE_GRAPH = "/reports/candidate_graph",
  GROUP = "/users/groups/",
  PERMISSION = "/users/get_permission/",
  VENDORS = "/vendors/vendor/",
  VENDOR_LIST = "/vendors/list/",
  CAMPAIGNS = "/schedulers/import_campaign_data/",
  CAMPAIGN_LIST = "/campaigns/campaign_list/",
  EMAIL_LIST_DETAILS = "/campaigns/campaign_email_list_data/",
  CAMPAIGN = "/campaigns/campaign/",
  CUSTOM_FIELDS = "/campaigns/custom_fields/",
  CAMPAIGN_DETAILS = "/campaigns/campaign_email_list_mapping_data/",
  IMPORT_CAMPAIGN_DATA = "/schedulers/import_campaign_data/",
  VENDOR_BULK_DELETE = "/vendors/vendor_bulk_delete/",
  VENDOR_LIST_DETAILS = "/vendors/list_data/",
  DESIGNATION = "/interviewers/designation/",
  INTERVIEWER = "/interviewers/interviewer/",
  TIMESLOT = "/interviewers/timeslot/",
  INTERVIEW = "/interviews/interview/",
  FEEDBACK = "/interviews/feedback/",
  SOURCE = "/interviews/source/",
  CANDIDATE = "/candidates/candidate/",
  ACTIVITY = "/candidates/activity/",
  RTR = "/candidates/rtr/",
  EMAIL_TEMPLATE = "/candidates/emailTemplate/",
  PLACEMENT = "/candidates/placement/",
  CLIENT = "/clients/client/",
  CANDIDATE_REPO = "/candidatesdocumentrepo/candidatesrepo/",
  CANDIDATE_STAGES = "/candidates/candidateStages/",
  CANDIDATE_STAGES_REVIEWS = "/candidates/candidateStageReviews/",
  IMPORT = "/candidates/import_candidates/",
  SENDMAIL = "/candidates/sendMail/",
  MAIL_VENDOR_LIST = "/vendors/send_vendor_list_email/",
  SEND_CAMPAIGN_EMAIL = "/campaigns/send_campaign_list_email/",
  IMPORT_CLIENT = "/clients/import_clients/",
  IMPORT_JD = "/jobdescriptions/import_jobdescription/",
  IMPORT_VENDOR = "/vendors/import_vendors/",
  EXPORT_CANDIDATE = "/candidates/export_candidate_model/",
  EXPORT_JD = "/jobdescriptions/export_jd_model/",
  UNASSIGNED_JOBS_STATUS = "/jobdescriptions/get_unassigned_jobs_status/",
  EXPORT_CLIENT = "/clients/export_client_model/",
  EXPORT_VENDORS = "/vendors/export_vendors_model/",
  VENDOR_TEMPLATE = "/vendors/emailTemplate/",
  CLIENT_DROPDOWN = "/reports/get_client_dropdown_list/",
  CLIENT_REVENUE_GRAPH = "/reports/get_client_revenue_graph/",
  CLIENT_REVENUE_CSV = "/reports/get_client_revenue_csv/",
  JD_DROPDOWN = "/reports/get_jobs_dropdown_list/",
  RECRUITER_JOBS = "/reports/get_jobs_dashboard_list/",
  REPO_BY_CANDIDATE_ID = "/candidatesdocumentrepo/get_repo/",
  CANDIDATE_INTERVIEW_LIST  = "/interviews/candidate_interviews_list/",
  EMAIL_CONGIF = "/vendors/emailConfiguration/",
  EMAIL_GET_CONFIG = "/vendors/get_config_obj/",
  GET_SUBMISSION_CANDIDATES = "/candidates/get_all/",
  GET_CANDIDATE_DROPDOWN = "/candidates/get_candidate_dropdown_list/",
  ASSIGNED_UNASSIGNED_JOBS = "/reports/get_assinged_dashboard_list/",
  GET_GRAPH_CANDIDATE_LIST = "/reports/get_graph_point_list",
  GET_CANDIDATE_JOB_STAGES = "/candidates/get_candidates_jobs_stages/",
  CANDIDATES_FOR_SUBMISSION_JOB = "/candidates/candidates_for_submission_job",
  GET_JOB_NAMES = "/schedulers/jobs",
  ADD_CANDIDATE_JOB_ACTIVITY = "/candidates/candidatesJobsStages/",
  ADD__INTERNAL_CANDIDATE_JOB_ACTIVITY = "/candidates/internalCandidateJobsStages/",
  DOWNLOAD_SPECIFIC_CANDIDATE_DETAILS = "/reports/export_graph_point_list/",
  GET_USER_DROPDOWN_LIST = "/jobdescriptions/get_user_dropdown_list/",
  DOWNLOAD_JOB_CANDIDATE_DETAILS = "/jobdescriptions/export_job_candidate_list/",
  GET_JOB_CURRENT_STATUS = "/jobdescriptions/get_job_current_status/",
  DOWNLOAD_ASSIGNMENT_HISTORY = "/jobdescriptions/export_assignment_history/",
  GET_DASHBOARD_SUMMARY = "/reports/get_dashboard_data/",
  GET_ANALYTICS = "/analytics/best_match_candidates/",

}